requirements.txt
openai
plaid-python
tiktoken
python-dotenv